refactor: Fix 2 'bugprone-signed-char-misuse' cases
Fixes 'bugprone-signed-char-misuse'

+++ src/base58.cpp
@@ -55,7 +55,7 @@ static const int8_t mapBase58[256] = {
     static_assert(std::size(mapBase58) == 256, "mapBase58.size() should be 256"); // guarantee not out of range
     while (*psz && !IsSpace(*psz)) {
         // Decode base58 character
-        int carry = mapBase58[(uint8_t)*psz];
+        int carry = static_cast<int>(mapBase58[(uint8_t)*psz]);
         if (carry == -1)  // Invalid b58 character
             return false;
         int i = 0;
